+++ tests/tests.js
@@ -2,7 +2,9 @@
   'use strict';
 
   var expect = chai.expect,
-      assert = chai.assert;
+      assert = chai.assert,
+      isPhantomJS = navigator.userAgent.indexOf('PhantomJS') === -1,
+      undefined;
 
   describe('Pablo test suite', function () {
     it('should load the Pablo library into the browser', function () {
@@ -67,6 +69,10 @@
 
       it('window.Document', function(){
         expect('Document' in window).to.equal(true);
+      });
+
+      it('window.HTMLDocument', function(){
+        expect('HTMLDocument' in window).to.equal(true);
       });
 
       it('svg.createSVGRect', function(){
@@ -1339,7 +1345,34 @@
           });
         });
 
-        describe('.transform()', function () {
+        describe.skip('.transition()', function () {
+          it('.transition(property) should add the CSS property name to the transition list', function(){
+              var subject = Pablo.circle().transition('opacity');
+              expect(subject.css('transition-property')).to.equal('opacity');
+          });
+
+          it('.transition(duration) should set the transition duration', function(){
+              var subject = Pablo.circle().transition(2000);
+              expect(subject.css('transition-duration')).to.equal('2s');
+          });
+
+          it('.transition(property, duration) should set the transition property and duration', function(){
+              var subject = Pablo.circle().transition('opacity', 2000);
+              expect(subject.css('transition-property')).to.equal('opacity');
+              expect(subject.css('transition-duration')).to.equal('2s');
+          });
+
+          it('.transition(properties, duration) should set the transition properties and duration', function(){
+              var subject = Pablo.circle().transition('opacity, stroke-width', 2000);
+              expect(subject.css('transition-property')).to.equal('opacity, stroke-width');
+              expect(subject.css('transition-duration')).to.equal('2s');
+          });
+        });
+
+        describe.skip('.transform()', function () {
+          var value = 'translate(50px) scale(0.5) rotate(45 50 50)';
+
+          /*
           it('.transform(functionName, value) should add a transform attribute and value to the element', function () {
             var subject = Pablo([Pablo.rect(), Pablo.rect()]);
 
@@ -1348,38 +1381,193 @@
             expect(subject[0].getAttribute('transform')).to.equal('rotate(45 50 50)');
             expect(subject[1].getAttribute('transform')).to.equal('rotate(45 50 50)');
           });
+*/
+
+          it('.transform() should return a key-value object of transforms for the first element', function () {
+            var subject = Pablo([Pablo.rect(), Pablo.rect()]),
+                result;
+
+            subject.eq(0).attr('transform', value);
+            result = subject.transform();
+
+            expect(result).to.be.an.object;
+            expect(result.scale).to.eql([0.5]);
+            expect(result.rotate).to.eql([45, 50, 50]);
+            expect(result.translate).to.eql(['50px']);
+          });
+
+          it('.transform(null) should clear the transform attribute for each element', function () {
+            var subject = Pablo([Pablo.rect(), Pablo.rect()]);
+
+            subject.attr('transform', value);
+            expect(subject[0].attr('transform')).to.equal(value);
+            expect(subject[1].attr('transform')).to.equal(value);
+
+            subject.transform(null);
+            expect(subject[0].attr('transform')).to.be.undefined;
+            expect(subject[1].attr('transform')).to.be.undefined;
+          });
+
+          // TODO: or use removeTransform ?
+          it('.transform(type, null) should clear that transform type for each element', function () {
+            var subject = Pablo([Pablo.rect(), Pablo.rect()]);
+
+            subject.attr('transform', value);
+            subject.transform('scale', null);
+            expect(subject[0].attr('transform').indexOf('scale')).to.equal(-1);
+            expect(subject[1].attr('transform').indexOf('scale')).to.equal(-1);
+
+            expect(subject[0].attr('transform').indexOf('rotate')).to.not.equal(-1);
+            expect(subject[1].attr('transform').indexOf('rotate')).to.not.equal(-1);
+            expect(subject[0].attr('transform').indexOf('translate')).to.not.equal(-1);
+            expect(subject[1].attr('transform').indexOf('translate')).to.not.equal(-1);
+          });
+
+          it('.transform(type) should return the value of the transform type for the first element', function () {
+            var subject = Pablo([Pablo.rect(), Pablo.rect()]);
+
+            subject.eq(0).attr('transform', value);
+            expect(subject.transform('scale')).to.eql([0.5]);
+            expect(subject.transform('rotate')).to.eql([45, 50, 50]);
+            expect(subject.transform('translate')).to.eql(['50px']);
+
+            subject.removeAttr('transform');
+            subject.eq(1).attr('transform', value);
+            expect(subject.transform('scale')).to.be.undefined;
+            expect(subject.transform('rotate')).to.be.undefined;
+            expect(subject.transform('translate')).to.be.undefined;
+          });
+
+          it('.transform(type, value) should set the simple value of the transform type for each element', function () {
+            var subject = Pablo([Pablo.rect(), Pablo.rect()]);
+
+            // Test with no transform applied
+            subject.transform('scale', 0.2);
+            expect(subject.eq(0).transform('scale')).to.eql([0.2]);
+            expect(subject.eq(1).transform('scale')).to.eql([0.2]);
+
+            subject.removeAttr('transform');
+            subject.attr('transform', value);
+
+            // Test with an existing transform applied
+            subject.transform('scale', 0.2);
+            expect(subject.eq(0).transform('scale')).to.eql([0.2]);
+            expect(subject.eq(1).transform('scale')).to.eql([0.2]);
+
+            expect(subject.eq(0).transform('rotate')).to.eql([45, 50, 50]);
+            expect(subject.eq(1).transform('rotate')).to.eql([45, 50, 50]);
+            expect(subject.eq(0).transform('translate')).to.eql(['50px']);
+            expect(subject.eq(1).transform('translate')).to.eql(['50px']);
+          });
+
+          it('.transform(type, value) should set the complex value of the transform type for each element', function () {
+            var subject = Pablo([Pablo.rect(), Pablo.rect()]),
+                expected = [0.5, 10, 10];
+
+            // Test with no transform applied
+            subject.transform('scale', expected);
+            expect(subject.eq(0).transform('scale')).to.eql(expected);
+            expect(subject.eq(1).transform('scale')).to.eql(expected);
+
+            subject.removeAttr('transform');
+            subject.attr('transform', value);
+
+            // Test with no transform applied
+            subject.transform('scale', expected);
+            expect(subject.eq(0).transform('scale')).to.eql(expected);
+            expect(subject.eq(1).transform('scale')).to.eql(expected);
+
+            expect(subject.eq(0).transform('rotate')).to.eql([45, 50, 50]);
+            expect(subject.eq(1).transform('rotate')).to.eql([45, 50, 50]);
+            expect(subject.eq(0).transform('translate')).to.eql(['50px']);
+            expect(subject.eq(1).transform('translate')).to.eql(['50px']);
+          });
+
+          // TODO: test attribute functions and values
+          // TODO: implement pluck('transform')
         });
 
         describe('.css()', function () {
-          it('.css(property) should return the specified css property of the element', function () {
-            expect(Pablo('#test-subjects').css('display')).to.equal('none');
+          describe('SVG elements', function(){
+            var subject = Pablo.circle({
+              style: 'opacity:0.5'
+            });
+
+            it('.css() should return the style object of the first element in the collection', function () {
+              var styleObj = subject.css(),
+                  result;
+
+              expect(styleObj).to.be.an.object;
+              result = Number(styleObj.opacity);
+              expect(result).to.equal(0.5);
+            });
+
+            // PhantomJS style object incomplete when run via grunt-contrib-mocha
+            if (isPhantomJS){
+              it('.css(property) should return the specified css property of the first element in the collection', function () {
+                var result = Number(subject.css('opacity'));
+                expect(result).to.equal(0.5);
+              });
+
+              it('.css(property, value) should set the specified css property of the first element in the collection', function () {
+                var result;
+
+                subject.css('stroke-width', '5px');
+                expect(subject.css('stroke-width')).to.equal('5px');
+              });
+
+              it('.css(styles) should set the specified css properties of the element in relation to the styles map', function () {
+                var subject = Pablo.circle().css({
+                      opacity: 0.2,
+                      'stroke-width': '10px'
+                    });
+
+                expect(Number(subject.css('opacity'))).to.equal(0.2);
+                expect(subject.css('stroke-width')).to.equal('10px');
+              });
+            }
           });
 
-          it('.css(property, value) should set the specified css property of the element', function () {
-            var subject = Pablo('#test-subjects');
+          describe('HTML elements', function(){
+            it('.css() should return the style object of the first element in the collection', function () {
+              var styleObj = Pablo('#test-subjects').css();
+              expect(styleObj).to.be.an.object;
+              expect(styleObj.display).to.equal('none');
+            });
 
-            subject.css('font-size', '20px');
+            // PhantomJS style object incomplete when run via grunt-contrib-mocha
+            if (isPhantomJS){
+              it('.css(property) should return the specified css property of the first element in the collection', function () {
+                expect(Pablo('#test-subjects').css('display')).to.equal('none');
+              });
 
-            expect(subject.css('font-size')).to.equal('20px');
-            resetTestSubjectStyles();
-          });
+              it('.css(property, value) should set the specified css property of the first element in the collection', function () {
+                var subject = Pablo('#test-subjects');
 
-          it('.css(styles) should set the specified css properties of the element in relation to the styles map', function () {
-            var subject = Pablo('#test-subjects').css({
-                  'font-size':   '20px',
-                  'font-weight': 'bold'
-                }),
-                result1 = subject.css('font-weight'),
-                result2 = subject.css('font-size');
+                subject.css('font-size', '20px');
+                expect(subject.css('font-size')).to.equal('20px');
+                resetTestSubjectStyles();
+              });
 
-            expect(result1).to.be.a('string');
-            expect(result1.length > 0).to.equal(true);
-            expect(result2).to.be.a('string');
-            expect(result2.length > 0).to.equal(true);
-            resetTestSubjectStyles();
+              it('.css(styles) should set the specified css properties of the element in relation to the styles map', function () {
+                var subject = Pablo('#test-subjects').css({
+                      'font-size':   '20px',
+                      'font-weight': 'bold'
+                    }),
+                    result1 = subject.css('font-weight'),
+                    result2 = subject.css('font-size');
+
+                expect(result1).to.be.a('string');
+                expect(result1.length > 0).to.equal(true);
+                expect(result2).to.be.a('string');
+                expect(result2.length > 0).to.equal(true);
+                resetTestSubjectStyles();
+              });
+            }
           });
         });
 
+        /*
         describe('.cssPrefix()', function () {
           it('.cssPrefix(prop, val) should set and prefix the passed css property with a browser namespace', function () {
             var subject = Pablo('#test-subjects'),
@@ -1432,6 +1620,7 @@
             expect(appliedWithPrefix2).to.equal(true);
           });
         });
+        */
 
         describe('.getValue()', function () {
           it('.getValue()', function () {
@@ -1540,7 +1729,7 @@
                 asArray     = subject.toArray();
 
             expect(asArray instanceof Array).to.equal(true);
-            expect(asArray.css).to.equal(undefined);
+            expect(asArray.css).to.be.undefined;
           });
         });
 
@@ -1862,9 +2051,9 @@
         describe('.pluck()', function () {
           it('.pluck(property) should return an array of values of default type attributes for each element in the collection', function () {
             var subject = Pablo([
-                                Pablo.rect({foo: '123'}),
-                                Pablo.ellipse({foo: '456'})
-                              ]),
+                  Pablo.rect({foo: '123'}),
+                  Pablo.ellipse({foo: '456'})
+                ]),
                 arr = subject.pluck('foo');
 
             expect(arr[0]).to.equal('123');
@@ -1873,9 +2062,9 @@
 
           it('.pluck(property, [attr]) should return an array of values of the type attributes for each element in the collection', function () {
             var subject = Pablo([
-                                Pablo.rect({foo: '123'}),
-                                Pablo.ellipse({foo: '456'})
-                              ]),
+                  Pablo.rect({foo: '123'}),
+                  Pablo.ellipse({foo: '456'})
+                ]),
                 arr = subject.pluck('foo','attr');
 
             expect(arr[0]).to.equal('123');
@@ -1908,19 +2097,23 @@
             expect(arr[1]).to.equal('456');
           });
 
-          it('.pluck(property, [css]) like above but with the type being a css rule', function () {
-            var subject = Pablo([
-                  Pablo.g().css({display: 'block'}),
-                  Pablo.g().css({display: 'inline'})
-                ]),
-                arr;
+          // PhantomJS style object incomplete when run via grunt-contrib-mocha
+          if (isPhantomJS){
+            it('.pluck(property, [css]) like above but with the type being a css rule', function () {
+              var subject = Pablo([
+                    Pablo.g().css({opacity: 0.5}),
+                    Pablo.g().css({opacity: 0.25})
+                  ]),
+                  arr;
 
-            arr = subject.pluck('display', 'css');
+              arr = subject.pluck('opacity', 'css');
 
-            expect(arr[0]).to.equal('block');
-            expect(arr[1]).to.equal('inline');
-          });
+              expect(arr[0]).to.equal('0.5');
+              expect(arr[1]).to.equal('0.25');
+            });
+          }
 
+          /*
           it('.pluck(property, [cssPrefix]) like above but with the type being a prefixed css rule', function () {
             var subject = Pablo([Pablo.rect(), Pablo.ellipse()]),
                 arr;
@@ -1935,6 +2128,7 @@
             expect(arr[0].indexOf('opacity 0.1s')).to.equal(0);
             expect(arr[1].indexOf('opacity 1s')).to.equal(0);
           });
+          */
         });
 
         describe('.select()', function () {
@@ -2106,7 +2300,7 @@
               expect(clone instanceof Pablo.Collection).to.equal(true);
               expect(clone[0] instanceof SVGRectElement).to.equal(true);
               expect(clone[0].getAttribute('foo')).to.equal('bar');
-              expect(clone.data('foo')).to.equal(undefined);
+              expect(clone.data('foo')).to.be.undefined;
             });
           });
 
@@ -2437,7 +2631,7 @@
 
         describe('.isDocument()', function () {
           it('.isDocument(obj) should return true if the HTML document object is passed', function () {
-            expect(Pablo.isDocument(document)).to.equal(true);
+            expect(Pablo.isDocument(document, true)).to.equal(true);
           });
 
           it('.isDocument(obj) should return false if a NodeList is passed', function () {
@@ -2487,10 +2681,15 @@
           });
         });
         
-        describe('.hyphensToCamelCase()', function () {
-          it('.hyphensToCamelCase() should return a camel cased string based of the passed hyphenated string', function () {
-            var cc = Pablo.hyphensToCamelCase('water-the-plants');
-            expect(cc).to.equal('waterThePlants');
+        describe('.camelCase()', function () {
+          it('.camelCase(str) should return a camelCase string based of the passed hyphenated string', function () {
+            expect(Pablo.camelCase('-moz-transition')).to.equal('mozTransition');
+            expect(Pablo.camelCase('moz-transition')).to.equal('mozTransition');
+          });
+
+          it('.camelCase(str, true) should return a camelCase string, with the first letter capitalised', function () {
+            expect(Pablo.camelCase('-moz-transition', true)).to.equal('MozTransition');
+            expect(Pablo.camelCase('moz-transition', true)).to.equal('MozTransition');
           });
         });
       });
@@ -2504,7 +2703,7 @@
         expect(subject.data('foo')).to.equal('bar');
 
         subject.removeData('foo');
-        expect(subject.data('foo')).to.equal(undefined);
+        expect(subject.data('foo')).to.be.undefined;
 
         subject.data('foo', 'bar');
         expect(subject.data('foo')).to.equal('bar');
@@ -2558,8 +2757,8 @@
 
           subject.each(function(el){
             var collection = Pablo(el);
-            expect(collection.data('foo')).to.equal(undefined);
-            expect(collection.data('fiz')).to.equal(undefined);
+            expect(collection.data('foo')).to.be.undefined;
+            expect(collection.data('fiz')).to.be.undefined;
           });
         });
 
@@ -2573,7 +2772,7 @@
 
           subject.removeData('foo');
 
-          expect(subject.data('foo')).to.equal(undefined);
+          expect(subject.data('foo')).to.be.undefined;
           expect(subject.data('fiz')).to.equal(123);
         });
       });
@@ -2603,7 +2802,7 @@
 
         it('when all of PabloCollection\'s data key/value pair are removed it should remove the data from the cache and the unique id for that PabloCollection', function () {
           subject.removeData();
-          expect(Pablo.cache[lastKey]).to.equal(undefined);
+          expect(Pablo.cache[lastKey]).to.be.undefined;
         });
       });
 
@@ -2648,11 +2847,11 @@
 
           subject.remove();
 
-          expect(subject.data('foo')).to.equal(undefined);
-          expect(subject.children().eq(0).data('foo')).to.equal(undefined);
-          expect(subject.children().eq(1).data('foo')).to.equal(undefined);
-          expect(subject.children().eq(0).firstChild().data('foo')).to.equal(undefined);
-          expect(subject.children().eq(1).firstChild().data('foo')).to.equal(undefined);
+          expect(subject.data('foo')).to.be.undefined;
+          expect(subject.children().eq(0).data('foo')).to.be.undefined;
+          expect(subject.children().eq(1).data('foo')).to.be.undefined;
+          expect(subject.children().eq(0).firstChild().data('foo')).to.be.undefined;
+          expect(subject.children().eq(1).firstChild().data('foo')).to.be.undefined;
         });
       });
 
@@ -2678,10 +2877,10 @@
           subject.empty();
 
           expect(subject.data('foo')).to.equal('bar');
-          expect(subject.children().eq(0).data('foo')).to.equal(undefined);
-          expect(subject.children().eq(1).data('foo')).to.equal(undefined);
-          expect(subject.children().eq(0).firstChild().data('foo')).to.equal(undefined);
-          expect(subject.children().eq(1).firstChild().data('foo')).to.equal(undefined);
+          expect(subject.children().eq(0).data('foo')).to.be.undefined;
+          expect(subject.children().eq(1).data('foo')).to.be.undefined;
+          expect(subject.children().eq(0).firstChild().data('foo')).to.be.undefined;
+          expect(subject.children().eq(1).firstChild().data('foo')).to.be.undefined;
         });
       });
     });
@@ -3434,13 +3633,15 @@
                 expect(img[0].height).to.equal(1);
               }
               else {
-                done(new Error('Incorrect `this` context'))
+                throw 'Incorrect `this` context';
               }
+              img.detach();
 
               if (count === 2){
                 done();
               }
-            });
+
+            }).appendTo('body');
           }
           testImage(subject1);
           testImage(subject2);
@@ -3450,9 +3651,12 @@
           function(done){
             subject1.toImage(function(img){
               var dataUrl = img.attr('src');
+
               expect(dataUrl.indexOf('data:image/svg+xml')).to.equal(0);
+
+              img.detach();
               done();
-            });
+            }).appendTo('body');
           }
         );
       }