+++ HISTORY.md
@@ -1,3 +1,8 @@
+## 0.3.7
+* css() now autodetect's browser vendor-prefixes
+* Removed cssPrefix()
+* Added `Pablo.userAgent`
+* Renamed `Pablo.v` to `Pablo.version`
 
 ## 0.3.6
 * Escape Unicode on data URL creation
@@ -16,7 +21,7 @@
 
 
 ## 0.3.3
-* Add bbox(), crop(), toImage(), toCanvas(), download()
+* Add bbox(), crop(), toImage(), toCanvas(), download(), stagger()
 * Add `Pablo.support` for fine-grained environment capabilities detection
 * Expose methods useful for advanced use or plugin writers
 * Use native `DOMParser` / `XMLSerializer` under the hood for creation from markup and outputting to markup